requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
coreapi==2.3.3
coreschema==0.0.4

Django==3.2.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
jwt==1.3.1
MarkupSafe==2.0.1
openapi-codec==1.3.2
pycparser==2.21
PyJWT==2.3.0
pytz==2021.3
requests==2.27.1
simplejson==3.17.6
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
cryptography==36.0.1
